requirements.txt
yfinance>=0.2.33
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
pytz>=2024.1
python-dotenv>=1.0.0
schedule>=1.2.0
